refactor: reduce duplication and use shared themes
- Use base adapter and DiffCallback for same logic
- Reduce kotlinx kover threshold
- Update testing
