+++ core/designsystem/src/main/res/layout/item_reveal_start.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
+  android:layout_width="wrap_content"
+  android:layout_height="wrap_content">
+
+  <com.google.android.material.listitem.ListItemRevealLayout
+    android:id="@+id/reveal_layout_start"
+    style="@style/ListItemRevealLayout.Start">
+
+    <Button
+      android:id="@+id/btn_delete"
+      style="@style/ListItemReveal.Button.Start"
+      android:contentDescription="@string/remove_from_watchlist" />
+
+  </com.google.android.material.listitem.ListItemRevealLayout>
+</androidx.coordinatorlayout.widget.CoordinatorLayout>
