+++ core/favoritewatchlist/src/test/kotlin/com/waffiq/bazz_movies/core/favoritewatchlist/ui/adapter/paging/FavoritePagingAdapterTest.kt
@@ -1,88 +1,43 @@
-package com.waffiq.bazz_movies.core.favoritewatchlist.ui.adapter
+package com.waffiq.bazz_movies.core.favoritewatchlist.ui.adapter.paging
 
-import android.content.Context
-import android.os.Looper
-import android.view.LayoutInflater
-import android.view.View
 import android.widget.FrameLayout
 import androidx.paging.PagingData
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
-import androidx.test.core.app.ApplicationProvider
 import com.google.android.material.listitem.SwipeableListItem
 import com.waffiq.bazz_movies.core.common.utils.Constants.MOVIE_MEDIA_TYPE
 import com.waffiq.bazz_movies.core.designsystem.R.id.container_result
 import com.waffiq.bazz_movies.core.designsystem.R.id.reveal_layout_end
 import com.waffiq.bazz_movies.core.designsystem.R.id.reveal_layout_start
-import com.waffiq.bazz_movies.core.designsystem.R.style.Base_Theme_BAZZ_movies
 import com.waffiq.bazz_movies.core.designsystem.databinding.ItemPagingFavoriteBinding
 import com.waffiq.bazz_movies.core.domain.MediaItem
+import com.waffiq.bazz_movies.core.favoritewatchlist.testutils.BaseAdapterPagingTest
 import com.waffiq.bazz_movies.core.favoritewatchlist.testutils.Constants.MOVIE_ORIGINAL_NAME
 import com.waffiq.bazz_movies.core.favoritewatchlist.testutils.Constants.MOVIE_ORIGINAL_TITLE
 import com.waffiq.bazz_movies.core.favoritewatchlist.testutils.Constants.MOVIE_TITLE
 import com.waffiq.bazz_movies.core.favoritewatchlist.testutils.Constants.TEST_DATE
 import com.waffiq.bazz_movies.core.favoritewatchlist.testutils.Constants.TEST_DATE_FORMATTED
-import com.waffiq.bazz_movies.core.test.MainDispatcherRule
-import com.waffiq.bazz_movies.navigation.INavigator
-import io.mockk.mockk
 import io.mockk.verify
 import junit.framework.TestCase.assertEquals
-import junit.framework.TestCase.assertFalse
 import junit.framework.TestCase.assertNotNull
 import junit.framework.TestCase.assertSame
-import junit.framework.TestCase.assertTrue
 import kotlinx.coroutines.test.advanceUntilIdle
 import kotlinx.coroutines.test.runTest
 import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
-import org.junit.runner.RunWith
-import org.robolectric.RobolectricTestRunner
-import org.robolectric.Shadows.shadowOf
 
-@RunWith(RobolectricTestRunner::class)
-class FavoritePagingAdapterTest {
-  private lateinit var context: Context
-  private lateinit var navigator: INavigator
+class FavoritePagingAdapterTest : BaseAdapterPagingTest() {
   private lateinit var adapter: FavoritePagingAdapter
-  private lateinit var inflater: LayoutInflater
   private lateinit var binding: ItemPagingFavoriteBinding
   private lateinit var viewHolder: FavoritePagingAdapter.ViewHolder
-  private lateinit var onDelete: (MediaItem, Int) -> Unit
-  private lateinit var onAddToWatchlist: (MediaItem, Int) -> Unit
-
-  @get:Rule
-  val mainDispatcherRule = MainDispatcherRule()
-
-  private val movieData = MediaItem(
-    mediaType = "movie",
-    name = "Test Movie Name",
-    title = MOVIE_TITLE,
-    originalTitle = MOVIE_ORIGINAL_TITLE,
-    originalName = MOVIE_ORIGINAL_NAME,
-    firstAirDate = TEST_DATE,
-    releaseDate = TEST_DATE,
-    listGenreIds = listOf(12),
-    voteAverage = 10f,
-    posterPath = "posterPath.jpg"
-  )
-  private val pagingData = PagingData.from(listOf(movieData))
 
   @Before
   fun setup() {
-    navigator = mockk(relaxed = true)
-    onDelete = mockk(relaxed = true)
-    onAddToWatchlist = mockk(relaxed = true)
+    super.setUp()
     adapter = FavoritePagingAdapter(
       navigator,
       MOVIE_MEDIA_TYPE,
       onDelete,
       onAddToWatchlist,
     )
-    context = ApplicationProvider.getApplicationContext<Context>().apply {
-      setTheme(Base_Theme_BAZZ_movies) // set the theme
-    }
-    inflater = LayoutInflater.from(context)
     binding = ItemPagingFavoriteBinding.inflate(inflater, null, false)
     viewHolder = adapter.ViewHolder(binding)
   }
@@ -186,13 +141,7 @@ class FavoritePagingAdapterTest {
   @Test
   fun onBindViewHolder_allAttributeIsNull_bindsCorrectMovieData() = runTest {
     // all null
-    adapter.submitData(
-      PagingData.from(
-        listOf(
-          MediaItem()
-        )
-      )
-    )
+    adapter.submitData(PagingData.from(listOf(MediaItem())))
     adapter.onBindViewHolder(viewHolder, 0)
 
     assertEquals("N/A", binding.tvTitle.text.toString())
@@ -215,7 +164,6 @@ class FavoritePagingAdapterTest {
 
   @Test
   fun onClick_whenClicked_opensMovieDetails() = runTest {
-    val pagingData = PagingData.from(listOf(movieData))
     adapter.submitData(pagingData)
     advanceUntilIdle()
 
@@ -227,25 +175,10 @@ class FavoritePagingAdapterTest {
 
   @Test
   fun bind_whenSwipedRight_callsOnDelete() = runTest {
-    val recyclerView = RecyclerView(context)
-    recyclerView.adapter = adapter
-    recyclerView.layoutManager = LinearLayoutManager(context)
+    provideRecyclerView(adapter) { adapter.submitData(pagingData) }
+    val viewHolder = provideViewHolder<FavoritePagingAdapter.ViewHolder>()
 
-    adapter.submitData(pagingData)
-    advanceUntilIdle()
-
-    // measure the RecyclerView
-    recyclerView.measure(
-      View.MeasureSpec.makeMeasureSpec(1080, View.MeasureSpec.EXACTLY),
-      View.MeasureSpec.makeMeasureSpec(1920, View.MeasureSpec.EXACTLY)
-    )
-    recyclerView.layout(0, 0, 1080, 1920)
-    shadowOf(Looper.getMainLooper()).idle()
-
-    val viewHolder =
-      recyclerView.findViewHolderForAdapterPosition(0) as? FavoritePagingAdapter.ViewHolder
     assertNotNull(viewHolder)
-
     assertNotNull(viewHolder!!.swipeCallback)
 
     // Trigger the actual callback
@@ -260,25 +193,10 @@ class FavoritePagingAdapterTest {
 
   @Test
   fun bind_whenSwipedLeft_callsOnAddToWatchlist() = runTest {
-    val recyclerView = RecyclerView(context)
-    recyclerView.adapter = adapter
-    recyclerView.layoutManager = LinearLayoutManager(context)
+    provideRecyclerView(adapter) { adapter.submitData(pagingData) }
+    val viewHolder = provideViewHolder<FavoritePagingAdapter.ViewHolder>()
 
-    adapter.submitData(pagingData)
-    advanceUntilIdle()
-
-    // measure the RecyclerView
-    recyclerView.measure(
-      View.MeasureSpec.makeMeasureSpec(1080, View.MeasureSpec.EXACTLY),
-      View.MeasureSpec.makeMeasureSpec(1920, View.MeasureSpec.EXACTLY)
-    )
-    recyclerView.layout(0, 0, 1080, 1920)
-    shadowOf(Looper.getMainLooper()).idle()
-
-    val viewHolder =
-      recyclerView.findViewHolderForAdapterPosition(0) as? FavoritePagingAdapter.ViewHolder
     assertNotNull(viewHolder)
-
     assertNotNull(viewHolder!!.swipeCallback)
 
     // Trigger the actual callback
@@ -293,25 +211,10 @@ class FavoritePagingAdapterTest {
 
   @Test
   fun bind_whenSwiped_othersState() = runTest {
-    val recyclerView = RecyclerView(context)
-    recyclerView.adapter = adapter
-    recyclerView.layoutManager = LinearLayoutManager(context)
+    provideRecyclerView(adapter) { adapter.submitData(pagingData) }
+    val viewHolder = provideViewHolder<FavoritePagingAdapter.ViewHolder>()
 
-    adapter.submitData(pagingData)
-    advanceUntilIdle()
-
-    // measure the RecyclerView
-    recyclerView.measure(
-      View.MeasureSpec.makeMeasureSpec(1080, View.MeasureSpec.EXACTLY),
-      View.MeasureSpec.makeMeasureSpec(1920, View.MeasureSpec.EXACTLY)
-    )
-    recyclerView.layout(0, 0, 1080, 1920)
-    shadowOf(Looper.getMainLooper()).idle()
-
-    val viewHolder =
-      recyclerView.findViewHolderForAdapterPosition(0) as? FavoritePagingAdapter.ViewHolder
     assertNotNull(viewHolder)
-
     assertNotNull(viewHolder!!.swipeCallback)
 
     // simulate dragging
@@ -327,37 +230,5 @@ class FavoritePagingAdapterTest {
       viewHolder.itemView.findViewById(container_result),
       100
     )
-  }
-
-  @Test
-  fun areItemsTheSame_whenIdAndMediaTypeIsSame_returnsTrue() {
-    val oldItem = MediaItem(id = 1, mediaType = "movie")
-    val newItem = MediaItem(id = 1, mediaType = "movie")
-
-    assertTrue(FavoritePagingAdapter.DIFF_CALLBACK.areItemsTheSame(oldItem, newItem))
-  }
-
-  @Test
-  fun areItemsTheSame_whenDifferentId_returnsFalse() {
-    val oldItem = MediaItem(id = 1, mediaType = "movie")
-    val newItem1 = MediaItem(id = 2, mediaType = "movie") // different ID
-
-    assertFalse(FavoritePagingAdapter.DIFF_CALLBACK.areItemsTheSame(oldItem, newItem1))
-  }
-
-  @Test
-  fun areContentsTheSame_whenIdAndMediaTypeIsSame_returnsTrue() {
-    val oldItem = MediaItem(id = 1, mediaType = "movie", title = "Movie 1")
-    val newItem = MediaItem(id = 1, mediaType = "movie", title = "Different Title")
-
-    assertTrue(FavoritePagingAdapter.DIFF_CALLBACK.areContentsTheSame(oldItem, newItem))
-  }
-
-  @Test
-  fun areContentsTheSame_whenDifferentId_returnsFalse() {
-    val oldItem = MediaItem(id = 1, mediaType = "movie")
-    val newItem1 = MediaItem(id = 2, mediaType = "movie") // different ID
-
-    assertFalse(FavoritePagingAdapter.DIFF_CALLBACK.areContentsTheSame(oldItem, newItem1))
   }
 }
