+++ feature/detail/src/test/kotlin/com/waffiq/bazz_movies/feature/detail/ui/adapter/CastAdapterTest.kt
@@ -34,7 +34,7 @@ class CastAdapterTest : BaseAdapterTest() {
   }
 
   @Test
-  fun sesubmitList_whenCalled_updatesListAndNotifiesChanges() {
+  fun submitList_whenCalled_updatesListAndNotifiesChanges() {
     val oldList = listOf(MediaCastItem(id = 1, name = "Old Cast"))
     val newList = listOf(
       MediaCastItem(id = 2, name = "New Cast"),
